+++ src/main/java/ru/obukhov/trader/market/model/PlacedMarketOrder.java
@@ -1,14 +1,18 @@
 
 package ru.obukhov.trader.market.model;
 
+import ru.tinkoff.piapi.contract.v1.OperationType;
+
+import java.math.BigDecimal;
+
 public record PlacedMarketOrder(
         String orderId,
         OperationType operation,
         OrderStatus status,
         String rejectReason,
         String message,
-        Integer requestedLots,
-        Integer executedLots,
+        BigDecimal requestedLots,
+        BigDecimal executedLots,
         MoneyAmount commission
 ) {
 }
